Delete User and loading Icon

+++ vue/src/components/AccountDetails.vue
@@ -1,6 +1,7 @@
 <template>
   <div class="account-details" >
     <h1>Account Details</h1>
+    <loading class="loading" v-if="isLoading"/>
     <div id="user-account" v-if="$store.state.user.role == 'user'">
         
         <h3>Placeholder ProfileName</h3>
@@ -13,7 +14,7 @@
       
     </div>
   <div> 
-    <button>Delete Account</button>
+    <button v-on:click="DeleteActiveUser">Delete Account</button>
     </div>
         
   </div>
@@ -21,11 +22,15 @@
 
 <script>
 import userService from "../services/UserService";
-export default {
 
+import loading from '../components/Loading';
+export default {
+  components: { loading },
   data() {
     return{
-      currentUser : {}
+      currentUser : {},
+      isLoading: false
+
     } 
   },
   created(){
@@ -34,8 +39,10 @@ export default {
   
   methods: {
     getActiveUser() {
+      this.isLoading = true;
       userService.getUser(this.$store.state.user.userId).then(response => {
         this.currentUser = response.data;
+        this.isLoading = false;
 
         if (response.status === 200 && response.data != null) {
           /* maybe send them somewhere? */
@@ -44,19 +51,25 @@ export default {
           /*this.$router.push(`/${name: login}`); */
         }
       });
+      
     },
       /* Delete Account Attempt  */
       DeleteActiveUser() {
-      userService.deleteUser(this.$store.state.user.userId).then(response => {
-        this.currentUser = response.data;
-        
-        if (response.status === 200 && response.data.length > 0) {
+      const verification = confirm("Are you sure you want to delete your account? Press OK to proceed.")
+      if(verification){
+              this.isLoading = true;
+      userService.deleteUser(this.$store.state.user.userId).then(response => {        if (response.status === 204) {
           /* maybe send them somewhere? */
+          alert("Account deleted successfully")
+          this.$store.commit("LOGOUT")
+          this.$router.push({ name: "login"})
         } else {
           alert("Account not found, please attempt to sign in again.")
           /*this.$router.push(`/${name: login}`); */
         }
-      });
+       });
+
+      }
     }
   }
 }
